+++ src/fr/iutvalence/info/dut/m2107/entities/Slime.java
@@ -15,7 +15,7 @@ public class Slime extends TerrestrialCreature {
 	
 	private short jumpWaitTime = 2000;
 	
-	public Slime(Vector2f pos, float rot, EntitySprite spr, Collider col, Vector2f vel, float spd, int health,
+	public Slime(Vector2f pos, float rot, EntitySprite spr, Collider col, Vector2f vel, short spd, int health,
 			int jumpHeight) {
 		super(pos, rot, spr, col, vel, spd, health, jumpHeight);
 	}
@@ -24,8 +24,8 @@ public class Slime extends TerrestrialCreature {
 	public void update(Layer layer) {
 		if(this.isGrounded) {
 			float moveX = GameWorld.player.pos.x - this.pos.x;
-			if(moveX < 0) this.spd = -Maths.fastAbs(this.spd);
-			else if(moveX > 0) this.spd = Maths.fastAbs(this.spd);
+			if(moveX < 0) this.spd = (short) -Maths.fastAbs(this.spd);
+			else if(moveX > 0) this.spd = (short) Maths.fastAbs(this.spd);
 		}
 		if(this.recoil == 0) {
 			if(!this.isGrounded) this.vel.x += this.spd;
