+++ src/fr/iutvalence/info/dut/m2107/entities/TerrestrialCreature.java
@@ -28,7 +28,7 @@ public class TerrestrialCreature extends LivingEntity {
 	protected boolean hasStepUp;
 	
 	
-	public TerrestrialCreature(Vector2f pos, float rot, EntitySprite spr, Collider col, Vector2f vel, float spd, int health,
+	public TerrestrialCreature(Vector2f pos, float rot, EntitySprite spr, Collider col, Vector2f vel, short spd, int health,
 						int jumpHeight) {
 		super(pos, rot, spr, col, vel, spd, health, jumpHeight);
 	}
